Correccion Superadmin/Noeliminar

+++ server.js
@@ -252,11 +252,16 @@ app.post("/api/admin/eliminar", auth, (req, res) => {
     return res.status(403).send("No autorizado");
   }
 
+  if (req.body.usuario === "Fonsecars") {
+    return res.status(400).send("No se puede eliminar el super administrador");
+  }
+
   db.prepare("DELETE FROM admin WHERE usuario = ?")
     .run(req.body.usuario);
 
   res.redirect("/admin.html");
 });
+
 
 // -------------------- SERVER --------------------
 app.get("/", (req, res) => {
